+++ kopych/src/hw3/Curs1/Teach1_1/Lek1_1/STUD1_1/hw1_1.java
@@ -1,4 +1,4 @@
-package Curs1.Teach1_1.Lek1_1.STUD1_1;
+package hw3.Curs1.Teach1_1.Lek1_1.STUD1_1;
 
 public class hw1_1 {
     public int id;
