+++ kopych/src/hw3/Curs1/Teach1_1/Lek1_1/STUD1_2/Stud1_2.java
@@ -1,9 +1,9 @@
-package Curs1.Teach1_1.Lek1_1.STUD1_2;
+package hw3.Curs1.Teach1_1.Lek1_1.STUD1_2;
 
-public class adhw1_2 {
+public class Stud1_2 {
     public int id;
     public static int counter = 1;
-    private adhw1_2(int id){
+    private Stud1_2(int id){
         this.id = id;
         counter ++;
     }
