+++ kopych/src/hw3/Curs1/Teach1_1/Lek1_1/STUD1_2/adhw1_2.java
@@ -1,9 +1,9 @@
-package Curs1.Teach1_2.Lek2_1.STUD2_2;
+package hw3.Curs1.Teach1_1.Lek1_1.STUD1_2;
 
-public class adhw2_2 {
+public class adhw1_2 {
     public int id;
     public static int counter = 1;
-    private adhw2_2(int id){
+    private adhw1_2(int id){
         this.id = id;
         counter ++;
     }
