+++ kopych/src/hw3/Curs1/Teach1_1/Lek1_1/STUD1_2/hw1_2.java
@@ -1,4 +1,4 @@
-package Curs1.Teach1_1.Lek1_1.STUD1_2;
+package hw3.Curs1.Teach1_1.Lek1_1.STUD1_2;
 
 public class hw1_2 {
     public int id;
