+++ kopych/src/hw3/Curs1/Teach1_1/Lek1_1/teach1_1.java
@@ -1,10 +1,10 @@
-package Curs1.Teach1_2.Lek2_1.STUD2_1;
+package hw3.Curs1.Teach1_1.Lek1_1;
 
-public class hw2_1 {
+public class teach1_1 {
     public int id;
     public static int counter = 1;
 
-    private hw2_1(int id) {
+    private teach1_1(int id) {
         this.id = id;
         counter++;
     }
