+++ kopych/src/hw3/Curs1/Teach1_2/Lek2_1/STUD2_1/hw2_1.java
@@ -1,10 +1,10 @@
-package Curs1.Teach1_2.Lek2_1.STUD2_2;
+package hw3.Curs1.Teach1_2.Lek2_1.STUD2_1;
 
-public class Stud2_2 {
+public class hw2_1 {
     public int id;
     public static int counter = 1;
 
-    private Stud2_2(int id) {
+    private hw2_1(int id) {
         this.id = id;
         counter++;
     }
