+++ kopych/src/hw3/Curs1/Teach1_2/Lek2_1/STUD2_2/adhw2_2.java
@@ -1,9 +1,9 @@
-package Curs1.Teach1_2.Lek2_1.STUD2_1;
+package hw3.Curs1.Teach1_2.Lek2_1.STUD2_2;
 
-public class adhw2_1 {
+public class adhw2_2 {
     public int id;
     public static int counter = 1;
-    private adhw2_1(int id){
+    private adhw2_2(int id){
         this.id = id;
         counter ++;
     }
