+++ kopych/src/hw3/Curs1/Teach1_2/Lek2_1/STUD2_2/hw2_2.java
@@ -1,10 +1,10 @@
-package Curs1.Teach1_2.Lek2_1;
+package hw3.Curs1.Teach1_2.Lek2_1.STUD2_2;
 
-public class teach1_2 {
+public class hw2_2 {
     public int id;
     public static int counter = 1;
 
-    private teach1_2(int id) {
+    private hw2_2(int id) {
         this.id = id;
         counter++;
     }
